Code review change - separate google map resource loading to a separate file
